Implemented Sparsifier Quality Checks and fixed issue with contract method
- Added method getFromAdj to Vertex
	- Gets the AdjNode containing a specific Vertex from the adjList
- Fixed issue with contract method see #19
- Removed code from Dijkstra's that set terminals
- Began implementing random terminal selection for quality checks after compression
- Updated timelog

+++ GraphCompression/src/graphUtils/General.java
@@ -26,9 +26,9 @@ public class General {
 			
 			for (int neighbIndex = 0; neighbIndex < chars.length; neighbIndex++) {	//Iterate over neighbouring vertices
 				Vertex currentVertex = newGraph.getVertex(vertIndex);
-				Float edgeWeight = Float.valueOf(chars[neighbIndex]);
+				double edgeWeight = Double.valueOf(chars[neighbIndex]);
 				
-				if (edgeWeight > 0.0) {
+				if (edgeWeight > 0) {
 					currentVertex.addToAdj(newGraph.getVertex(neighbIndex), edgeWeight);
 				}
 			}
